Allow cookie based authentication
This allows aurora client to connect to servers which are behind a
frontend which expects some sort of cookie to autheticate and authorize
users.

A cookie_jar option can be specified in the `~/.aurora/config.json`
to specify a file where the cookie jar is located. Such a cookie jar,
in MozillanCookieJar format, will be used to create the session and
therefore all the subsequent requests will use it.